Adding more testing scenearios for baskets

+++ src/main/java/bu/ist/visreg/basket/filesystem/FileBasket.java
@@ -64,7 +64,10 @@ public class FileBasket extends Basket {
 					subitem.persist();
 					addBasketItem(subitem);
 				}
-				bi.delete();
+				if(subitems.size() > 1) {
+					System.out.println("Deleting: " + bi.getPathname());
+					bi.delete();
+				}
 				break;
 			}
 		}
